typing: fix App's core API - App (#16949)

(cherry picked from commit 16e39a2407ac6695d8a66a6c91776200f4677ccb)

+++ src/lightning_app/core/queues.py
@@ -174,7 +174,7 @@ class BaseQueue(ABC):
         pass
 
     @abstractmethod
-    def get(self, timeout: Optional[int] = None) -> Any:
+    def get(self, timeout: Optional[float] = None) -> Any:
         """Returns the left most element of the queue.
 
         Parameters
@@ -380,7 +380,7 @@ class HTTPQueue(BaseQueue):
             return False
         return False
 
-    def get(self, timeout: Optional[int] = None) -> Any:
+    def get(self, timeout: Optional[float] = None) -> Any:
         if not self.app_id:
             raise ValueError(f"App ID couldn't be extracted from the queue name: {self.name}")
